refactor album package structure
