+++ src/main/java/com/ilnitsk/animusic/anime/Anime.java
@@ -2,7 +2,7 @@ package com.ilnitsk.animusic.anime;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
-import com.ilnitsk.animusic.album.Album;
+import com.ilnitsk.animusic.album.dao.Album;
 import com.ilnitsk.animusic.soundtrack.Soundtrack;
 import jakarta.persistence.*;
 import lombok.*;
