+++ src/test/java/com/ilnitsk/animusic/repositories/AlbumRepositoryIntegrationTest.java
@@ -1,9 +1,9 @@
 package com.ilnitsk.animusic.repositories;
 
 
-import com.ilnitsk.animusic.album.Album;
-import com.ilnitsk.animusic.album.AlbumRepository;
+import com.ilnitsk.animusic.album.dao.Album;
 import com.ilnitsk.animusic.album.dto.CreateAlbumRequest;
+import com.ilnitsk.animusic.album.repository.AlbumRepository;
 import com.ilnitsk.animusic.anime.Anime;
 import com.ilnitsk.animusic.anime.AnimeRepository;
 import com.ilnitsk.animusic.soundtrack.Soundtrack;
