+++ src/test/java/com/ilnitsk/animusic/repositories/AlbumRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.ilnitsk.animusic.repositories;
 
-import com.ilnitsk.animusic.album.Album;
-import com.ilnitsk.animusic.album.AlbumRepository;
+import com.ilnitsk.animusic.album.dao.Album;
+import com.ilnitsk.animusic.album.repository.AlbumRepository;
 import com.ilnitsk.animusic.anime.Anime;
 import com.ilnitsk.animusic.anime.AnimeRepository;
 import org.junit.jupiter.api.Test;
